moved getgroups, grouplist to aaa
